+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 ### 0.3.7 - Unreleased
+- Added new nft_chain, nft_contract, nft_token fields
 - Moved changelog
 - Updated examples